requirements.txt
manifest-ml[all]==0.1.8
dbt-duckdb==1.6.0
streamlit>=1.26.0